requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
python-dotenv
